+++ classes/Hook/HookActionOrderStatusPostUpdate.php
@@ -74,11 +74,7 @@ class HookActionOrderStatusPostUpdate implements HookInterface
         // If it was not already refunded
         if ($gaRefundSent === false) {
             // We refund it and set the "sent" flag to true
-            if ((bool) Configuration::get('GA_V4_ENABLED')) {
-                $js = $this->getGoogleAnalytics4($this->params['id_order']);
-            } else {
-                $js = $this->getUniversalAnalytics($this->params['id_order']);
-            }
+            $js = $this->getGoogleAnalytics4($this->params['id_order']);
             $this->context->cookie->__set('ga_admin_refund', $js);
             $this->context->cookie->write();
 
@@ -97,14 +93,6 @@ class HookActionOrderStatusPostUpdate implements HookInterface
     public function setParams($params)
     {
         $this->params = $params;
-    }
-
-    /**
-     * @param int $idOrder
-     */
-    protected function getUniversalAnalytics($idOrder)
-    {
-        return 'MBG.refundByOrderId(' . json_encode(['id' => $idOrder]) . ');';
     }
 
     /**
