+++ config.xml
@@ -2,7 +2,7 @@
 <module>
 	<name>ps_googleanalytics</name>
 	<displayName><![CDATA[Google Analytics]]></displayName>
-	<version><![CDATA[4.3.0]]></version>
+	<version><![CDATA[5.0.0]]></version>
 	<description><![CDATA[Gain clear insights into important metrics about your customers, using Google Analytics]]></description>
 	<author><![CDATA[PrestaShop]]></author>
 	<tab><![CDATA[analytics_stats]]></tab>
